Trying again with the player physics system, but this time much simpler
