+++ config.json
@@ -13,7 +13,7 @@
     "ground_color": [146,137,60],
     "ground_height": 50.0,
     "jump_button": "South",
-    "jump_force": 25.0,
+    "jump_force": 10.0,
     "life_image": "/heart.png",
     "life_width": 32,
     "map": [
@@ -34,7 +34,7 @@
     "player_standing_image_width": 160.0,
     "player_standing_image": "/jungle_player_standing.png",
     "player_starting_x": 150.0,
-    "player_starting_y": 675.0,
+    "player_starting_y": 710.0,
     "player_width": 80.0,
     "resolution_x": 1920,
     "resolution_y": 1080,
@@ -44,6 +44,7 @@
     "start_index": 0,
     "surface_bottom_height": 25.0,
     "surface_color": [189,179,94],
+    "surface_floor_y": 790,
     "surface_top_height": 125,
     "title_subtext": "Press start to begin",
     "title": "Jungle",
