+++ src/config/mod.rs
@@ -63,6 +63,7 @@ pub struct Config {
     pub surface_bottom_height: f32,
     #[serde(with = "crate::config::serde_color")]
     pub surface_color: Color,
+    pub surface_floor_y: f32,
     pub surface_top_height: f32,
     pub title_subtext: String,
     pub title: String,
@@ -95,7 +96,7 @@ impl Default for Config {
             ground_color: Color::from_rgb(146, 137, 60),
             ground_height: 50.0,
             jump_button: Button::South,
-            jump_force: 25.0,
+            jump_force: 1.0,
             life_image: "/heart.png".to_owned(),
             life_width: 32.0,
             map: vec![vec![MapFeature::Pit1]],
@@ -113,7 +114,7 @@ impl Default for Config {
             player_standing_image_width: 160.0,
             player_standing_image: "/jungle_player_standing.png".to_owned(),
             player_starting_x: 150.0,
-            player_starting_y: 675.0,
+            player_starting_y: 710.0,
             player_width: 80.0,
             resolution_x: 1920.0,
             resolution_y: 1080.0,
@@ -123,6 +124,7 @@ impl Default for Config {
             start_index: 0,
             surface_bottom_height: 25.0,
             surface_color: Color::from_rgb(189, 179, 94),
+            surface_floor_y: 790.0,
             surface_top_height: 125.0,
             title_subtext: "Press start to begin".to_owned(),
             title: "Jungle".to_owned(),
